refactor dashboard into hook and components
